refactor: _zlib 폴더 삭제
* 폴더 구조 단순화
